,の追加

+++ app/src/main/java/com/sotugyouseisaku/app/Record/CartViewRecord.java
@@ -8,7 +8,7 @@ public class CartViewRecord {
     private int productId;     // 商品ID
     private int quantity;      // 数量
     private String productName;// 商品名
-    private double price;         // 価格
+    private int price;         // 価格
     private String jyoutai;    // 商品状態（新品・中古）
     private String maker;      // メーカー名
     private String category;   // カテゴリ名
